Convert data to T (SI unit) before inversion (#68)
Currently, the data input for the inversion is in nanoteslas (nT). But
it should be in teslas (T), the proper SI unit, for calculations. This
lead to final results of the inversion being many orders of magnitude
bigger. That was addressed before by multiplying the estimates by `1e-9`
(value used to convert from nanotesla to tesla),

# Fixes  
- Convert data from `nT` to `T` before inversion
- Remove `temporay` folder