Refactor the linear inversion class, update docs
Refactor the `MagneticMomentBz` class to be more like the other Fatiando
inversion classes. Make the Jacobian calculation public, move some of
the coordinate conversions into it, organize the docstrings, cite
Gelson's paper, and improve description of inputs and outputs, use an
array to build the Jacobian instead of lists (still slow since it's
using a loop). Replaced the napoleon sphinx extension with numpydoc to
get better formatted API docs, which removes the need to list methods in
the docstring. Organize dependencies since I was editing pyproject.toml
anyway, moving ensaio to a docs building dependency instead of
a run-time dependency (it's not needed to run magali, only the examples
and tests).