Use our own input validation functions
The inversion is using Verde's functions to check if inputs are valid.
This will be a problem since these functions will be removed from Verde
2.0. It's also better to have our own since they can check for more
specific things. For example, we always need 3 coordinates but the Verde
functions accept any number. Create our own functions in new module
`_validation.py` and use them in the inversion class.