Add unit conversion functions instead of constants
Rename the `_constants.py` module to `_units.py` and create functions to
handle the unit conversions instead of multiplying by the constants in
other modules. These are relatively trivial but best to avoid dividing
or multiplying by mistake. These are still private functions, not part
of the public API. Rename the `_convert_micrometer_to_meter` function to
`coordinates_micrometers_to_meters` and move it to the new module.