Move to src layout (#86)
This is now recommended practice for Python packages. The package code
goes in a `src` folder and the tests are not part of the package anymore
and live in a `test` folder. An advantage of this is that we don't need
to make a temp folder to run the tests to make sure we're testing the
installed version of the package. Changed relative imports in tests to
absolute imports now that it's not a package. Took this chance to move
the coverage configuration to `pyproject.toml` and remove the
`_version.py` file which is no longer needed now that we don't host
datasets in Magali.

---------

Co-authored-by: Yago M. Castro <yagomcastro1@gmail.com>