Use numba to calculate the linear Jacobian matrix (#99)
Accelerate the Jacobian by jit-compiling with numba. Also start
returning the Jacobian in nT instead of T so we don't need to convert
the data units. Refactor the tests to avoid doing unnecessary things
like plotting and generating random directions. Includes a test for the
non-jit-compiled Jacobian as well using a simple forward modeling
comparison to make sure all the code is covered by the tests.

**Relevant issues/PRs:** Fixes #81