Add nonlinear inversion for dipole position and magnetic moment using Levenberg–Marquardt (#107)
This PR introduces a nonlinear inversion algorithm for estimating the 3D
location (x, y, z) and magnetic moment (mx, my, mz) of a single magnetic
dipole from vertical magnetic field measurements (Bz). The method is
implemented in the new NonlinearMagneticDipoleBz class and is based on
the Levenberg-Marquardt algorithm, offering improved stability and
convergence for nonlinear problems.

This implementation performs the inversion of both the dipole position
and magnetic moment using Bz data. It employs a two-loop optimization
strategy in which the outer loop handles the nonlinear update of the
dipole location, while the inner loop solves a linear system to estimate
the magnetic moment.

---------

Co-authored-by: Gelson F. Souza-Junior <gelson.ferreira@iag.usp.br>