Adapt forward modeling for multiple components
If we want to work with the diamond directions, we'll need to start
calculating the full magnetic field vector of the dipoles. Adapt the
functions to have a `field` argument like in Harmonica. Use xarray
directly to make the grid instead of Verde. Move forward modeling
functions to a `_forward.py` module.