Fix nonlinear iterative inversion removing modeled data from the entire map instead of locally (#128)
This PR fixes a logic issue in the iterative_nonlinear_inversion
implementation where the inversion algorithm incorrectly removes the
modeled data from only within the bounding boxes area instead of the
entire map. Unnecessary upward continuation was also applied to the
dataset.

Additionally, this PR adds a function to compute the angular distance
between two vectors, which is used in the unit tests to compare the
estimated dipole moment vectors with the true ones.This PR additionally
introduces a private helper function `_clip_step` to enforce a maximum
step length during the Levenberg–Marquardt parameter update.

**Relevant issues/PRs:**
#120